Demand a RoundTripper, not a http.Transport
As it turns out, http.Transport is an implementation of only one such
transport, a sort of okay one intended to be used for network traffic
(too bad it doesn't have an obvious place where one can introduce
timeouts).

One might expect that http.Transport should be the interface type, and
what it currently is should be named something else, but for whatever
reason things are not this way.

Discovered when trying to figure out how to implement client timeouts
implemented via connection SetDeadline, and realizing it seems as
though one will have to write an entire RoundTripper to do so cleanly.

Signed-off-by: Daniel Farina <daniel@fdr.io>