Rename Multi to "Client" and Client to "MiniClient"
The thinking here is that the Minimal Client might be useful to some
people, but most people are going to want Multi anyway, which is
capable of batching, dropping, and posting of its own volition.

So, with that thought in mind, rename Multi to be just "Client" so
that people don't write logplexc consumers that do normally
undesirable things, such as blocking in event of backlog.

Signed-off-by: Daniel Farina <daniel@heroku.com>