Support dropping and drop metrics
In all the shuffle it seems that buffer dropping has been forgotten:
previously, it could only be a side effect of PostMessages, which is
simply not called when one runs out of concurrency tokens.

Now that what was the "Multi" client is just "Client" and what was
"Client" is now "MiniClient" it seems less bad to expose some extra
complexity from MiniClient to emit Bundles of work (and their
accompanying statistics) that must be subsequently POSTed, or simply
thrown away to be dropped.

Signed-off-by: Daniel Farina <daniel@heroku.com>