Amendments to README.
Seems like it could be a test to prevent such regressions.  Maybe
later.

As a positive, the changes to rotate the old "Client" to "MiniClient"
and the old "Multi" to "Client" have made the example more automatic,
and thus smaller, except for the configuration of some new features
most people are going to want anyway.

Signed-off-by: Daniel Farina <daniel@heroku.com>