Add special case for TargetLogLatency == 0
A bit of a rough hack, but if such a latency is passed, bypass the
ticker goroutine altogether.

An alternative would be to tune the RequestSizeTrigger to zero, but
this is an exposed field, and modifying its value to not be that which
the caller has passed in the configuration seems to violate the
principle of least astonishment.

Signed-off-by: Daniel Farina <daniel@fdr.io>